Move integration trade test to separate file
